Add teams

+++ crm_django/lead/views.py
@@ -1,4 +1,4 @@
-from rest_framework import viewsets #, decorators, response
+from rest_framework import viewsets
 from .serializers import LeadSerializer
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
@@ -9,9 +9,12 @@ class LeadViewSet(viewsets.ModelViewSet):
     queryset = Lead.objects.all()
 
     def get_queryset(self):
-       return self.queryset.filter()
+       return self.queryset.filter(created_by=self.request.user)
+
+    def perform_create(self, serializer):
+        serializer.save(created_by=self.request.user)
 
 @api_view(['POST'])
 def delete_lead(request,lead_id):
-    Lead.objects.filter(id=lead_id).delete()
+    Lead.objects.filter(id=lead_id, created_by=request.user).delete()
     return Response({'message':'Deleted'})
